feat(rs-16): integrate 16-core digital twin and R-SPU 2.0 refactor
- Refactored R-SPU 2.0 to support 1024 registers (u16) and 64-bit instructions.
- Fixed a critical emitter bug in 'temporal.rs' that truncated delayed signals (_d1) to 1-bit.
- Implemented 'Robot Digital Twin' in MIRR (pendulum.mirr, controller.mirr).
- Integrated the digital twin with the full 16-core RS-16 design in 'rspu_top.mirr'.
- Updated physical models (regfile.sv, ram.sv) to match architectural expansion.
- Stabilized physics logic using module-level signals to satisfy the 64-bit width solver.
- Verified system stability via BMC/Z3 formal solver.

+++ src/cert/mod.rs
@@ -112,7 +112,7 @@ pub struct ProofCertificate {
 /// Build a proof certificate from a totality analysis result and program binary.
 pub fn build_certificate(
     totality: &crate::totality::TotalityResult,
-    program_binary: &[u32],
+    program_binary: &[u64],
     module: &crate::ast::program::Module,
 ) -> ProofCertificate {
     use crate::ast::types::SignalKind;
